Splitting the client libraries into separate files.
