+++ api_client/python/timesketch_api_client/sketch_test.py
@@ -1,0 +1,56 @@
+# Copyright 2017 Google Inc. All rights reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+"""Tests for the Timesketch API client"""
+from __future__ import unicode_literals
+
+import unittest
+import mock
+
+from . import client
+from . import test_lib
+from . import timeline as timeline_lib
+from . import view as view_lib
+
+
+class SketchTest(unittest.TestCase):
+    """Test Sketch object."""
+
+    @mock.patch('requests.Session', test_lib.mock_session)
+    def setUp(self):
+        """Setup test case."""
+        self.api_client = client.TimesketchApi(
+            'http://127.0.0.1', 'test', 'test')
+        self.sketch = self.api_client.get_sketch(1)
+
+    # TODO: Add test for upload()
+
+    def test_get_views(self):
+        """Test to get a view."""
+        views = self.sketch.list_views()
+        self.assertIsInstance(views, list)
+        self.assertEqual(len(views), 2)
+        self.assertIsInstance(views[0], view_lib.View)
+
+    def test_get_timelines(self):
+        """Test to get a timeline."""
+        timelines = self.sketch.list_timelines()
+        self.assertIsInstance(timelines, list)
+        self.assertEqual(len(timelines), 2)
+        self.assertIsInstance(timelines[0], timeline_lib.Timeline)
+
+    def test_explore(self):
+        """Tests to explore a timeline."""
+        results = self.sketch.explore(query_string='description:test')
+        self.assertEqual(len(results['objects']), 1)
+        self.assertIsInstance(results['objects'], list)
