+++ api_client/python/timesketch_api_client/view_test.py
@@ -18,37 +18,23 @@ import unittest
 import mock
 
 from . import client
-from . import sketch as sketch_lib
+from . import view as view_lib
 from . import test_lib
 
 
-class TimesketchApiTest(unittest.TestCase):
-    """Test TimesketchApi"""
+class ViewTest(unittest.TestCase):
+    """Test View object."""
 
     @mock.patch('requests.Session', test_lib.mock_session)
     def setUp(self):
         """Setup test case."""
         self.api_client = client.TimesketchApi(
             'http://127.0.0.1', 'test', 'test')
+        self.sketch = self.api_client.get_sketch(1)
 
-    def test_fetch_resource_data(self):
-        """Test fetch resource."""
-        response = self.api_client.fetch_resource_data('sketches/')
-        self.assertIsInstance(response, dict)
-
-    # TODO: Add test for create_sketch()
-
-    def test_get_sketch(self):
-        """Test to get a sketch."""
-        sketch = self.api_client.get_sketch(1)
-        self.assertIsInstance(sketch, sketch_lib.Sketch)
-        self.assertEqual(sketch.id, 1)
-        self.assertEqual(sketch.name, 'test')
-        self.assertEqual(sketch.description, 'test')
-
-    def test_get_sketches(self):
-        """Test to get a list of sketches."""
-        sketches = self.api_client.list_sketches()
-        self.assertIsInstance(sketches, list)
-        self.assertEqual(len(sketches), 1)
-        self.assertIsInstance(sketches[0], client.Sketch)
+    def test_view(self):
+        """Test View object."""
+        view = self.sketch.list_views()[0]
+        self.assertIsInstance(view, view_lib.View)
+        self.assertEqual(view.id, 1)
+        self.assertEqual(view.name, 'test')
